Complete the majority of the CreateCollabVoiceMemoCell. Only remaining features until VoiceMemos in collabs are fully functional are the saving of the voiceMemo's to Firebase, and the retrieval of them in the CollabViewController.

+++ Block Pro/Model/VoiceMemo.swift
@@ -14,7 +14,7 @@ struct VoiceMemo {
     
     var name: String?
     
-    var length: Int?
+    var length: Double?
     
     var dateCreated: Date?
 }
